Use different load-directory implementation
Using the previous one always printed a "wrong-type-argument
number-or-marker-p conses" error and froze emacs-nox for a few seconds
at startup. Which was annoying when using it as git's core editor.

I'm not an emacs guru to track this error, so instead I replaced the
old implementation with something else found on the internet and which
does not produce any errors or freezes.

My guess is that at some point Emacs incompatibly changed something an
old implementation used. Based on
http://lists.gnu.org/archive/html/bug-gnu-emacs/2012-12/msg01015.html.